+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Use async `Tuc.DomainResolver`
 
 ## 1.6.0 - 2021-08-12
 - Update dependencies